No hay cambios importantes en el funcionamiento, se añaden tests aleatorios.

+++ P4/head/hextree.h
@@ -2,7 +2,7 @@
 #define _HEXTREE_H
 
 struct hextree_node {
-	short value;
+	float value;
 	short index;
 	struct hextree_node *children[6];
 };
@@ -13,7 +13,7 @@ short addHextreeNode(struct hextree_node *parent, short position, short index);
 short isHextreeLeaf(struct hextree_node *node);
 struct hextree_node* getChildren(struct hextree_node *parent, short position);
 short getIndex(struct hextree_node *node);
-short setValue(struct hextree_node *node, short value);
+short setValue(struct hextree_node *node, float value);
 struct hextree_node* negaMax(struct hextree_node *node, short color);
 
 #endif
